Move to dedicated package and split into several files
